fix(ci): MSVC /W4 /WX 오류 수정 — C4505/C4365 다수 파일
gain_correct.cpp:
  - C4505: ulp_difference, check_parity, apply_gain_fma에 [[maybe_unused]] 추가
    (정의되었으나 호출되지 않는 static 함수)

offset_correct.cpp:
  - C4365: xpe_has_avx512f에서 xcr_mask & 0xE6 → 0xE6ULL
    (unsigned long long과 int 리터럴 혼용)

defect_correct.cpp:
  - C4365: positions.size() >= 2 → 2u
  - C4365: values.size() / 2 → 2u

ghost_correct.cpp:
  - C4365: (y + dy) * W + (x + dx) — uint32_t와 int 혼용 산술
    → static_cast<size_t>(static_cast<int>(y) + dy) * ... 로 수정

calibration_manager.cpp:
  - C4365: c & 1 → 1u
  - C4365: crc ^ data[i] → static_cast<uint32_t>(data[i]), & 0xFF → 0xFFu

rle_codec.cpp:
  - C4365: val & 0xFF → 0xFFu (write_u32_le, 4개소)
  - C4365: len < 1024, len * 5, 4096 → size_t{} 리터럴
  - C4365: pos + 5, pos + 1 → size_t{5}, size_t{1}
  - C4365: enc_len % 5 → size_t{5}, i += 5 → size_t{5}
  - C4365: encoded[i + 1] → i + size_t{1}

+++ modules/preprocess/src/ghost_correct.cpp
@@ -171,7 +171,7 @@ namespace {
                     int count = 0;
                     for (int dy = -1; dy <= 1; ++dy) {
                         for (int dx = -1; dx <= 1; ++dx) {
-                            const size_t ni = (y + dy) * W + (x + dx);
+                            const size_t ni = static_cast<size_t>(static_cast<int>(y) + dy) * static_cast<size_t>(W) + static_cast<size_t>(static_cast<int>(x) + dx);
                             if (ni < n && std::isfinite(px[ni])) {
                                 localMean += px[ni];
                                 ++count;
